se agrego abm usuarios e interaccion con jwt

+++ src/controllers/phones.controllers.js
@@ -75,7 +75,6 @@ phonesController.getPhones = async (req, res) => {
             mensaje: "error al obtener informacion",
         });
     }
-
 };
 phonesController.getPhonesJWT = async (req, res) => {
     try {
@@ -84,12 +83,12 @@ phonesController.getPhonesJWT = async (req, res) => {
             const baererToken = baererHeader.split(" ")[1]
             req.token = baererToken;
             const permission = await permissionJWTVerify(baererToken, JWTFlag.permissionNameView)
-            if (permission == false) {
+            if (permission.flag == false) {
                 res.sendStatus(403)
             } else {
                 try {
-                    if (req.query.all && req.query.all == "true") {
-                        if (req.query.page) {
+                    if (req.query.all && req.query.all == "true" && req.query.page) {
+                        if (permission.user.user.permisions.find(permissionsAux => permissionsAux === 'admin')) {
                             const page = req.query.page
                             let elements = 5
                             if (req.query.elements) {
@@ -103,40 +102,45 @@ phonesController.getPhonesJWT = async (req, res) => {
                             res.status(200).json(result);
                         } else {
                             res.status(500).json({
-                                mensaje: "error al obtener paginado",
+                                mensaje: "permisos insuficientes",
                             });
                         }
                     } else {
-                        if (req.query.bu && req.query.page) {
-                            const bu = req.query.bu
-                            const page = req.query.page
-                            let elements = 5
-                            let counter = 0
-                            const phones = await phone.find({ businessUnit: bu });
-                            if (req.query.elements) {
-                                elements = req.query.elements
-                                if (parseInt(req.query.elements) > phones.length) {
-                                    elements = phone.length
+                        if (permission.user.user.permisions.find(permissionsAux => permissionsAux === 'view') || permission.user.user.permisions.find(permissionsAux => permissionsAux === 'admin')) {
+                            if (req.query.bu && req.query.page) {
+                                const bu = req.query.bu
+                                const page = req.query.page
+                                let elements = 5
+                                if (req.query.elements) {
+                                    elements = req.query.elements
+                                }
+                                if (permission.user.user.permisions.find(permissionsAux => permissionsAux === 'admin')) {
+                                    const phonesCount = await phone.count({ businessUnit: bu });
+                                    const phones = await phone.find({ businessUnit: bu });
+                                    const result = []
+                                    result.push({ pagination: { "page": page, "maxObjectsPerPage": parseInt(elements), "totalObjects": phonesCount } })
+                                    result.push({ phones: phones })
+                                    res.status(200).json(result);
+                                } else {
+                                    const phonesCount = await phone.count({ businessUnit: bu, bussinesAccountId: permission.user.user.bussinesAccountId });
+                                    const phones = await phone.find({ businessUnit: bu, bussinesAccountId: permission.user.user.bussinesAccountId });
+                                    const result = []
+                                    result.push({ pagination: { "page": page, "maxObjectsPerPage": parseInt(elements), "totalObjects": phonesCount } })
+                                    result.push({ phones: phones })
+                                    res.status(200).json(result);
                                 }
                             }
-                            const result = []
-                            const phonesAux = []
-                            result.push({ pagination: { "page": page, "maxObjectsPerPage": parseInt(elements), "totalObjects": phones.length } })
-                            for (let index = (parseInt(page) - 1) * parseInt(elements); index < phones.length; index++) {
-                                counter = counter + 1
-                                if (counter > elements) {
-                                    break
-                                }
-                                phonesAux.push(phones[index])
+                            else {
+                                res.status(500).json({
+                                    mensaje: "error al obtener paginado",
+                                });
                             }
-                            result.push({ phones: phonesAux })
-                            res.status(200).json(result);
-                        }
-                        else {
+                        } else {
                             res.status(500).json({
-                                mensaje: "error al obtener paginado",
+                                mensaje: "permisos insuficientes",
                             });
                         }
+
                     }
 
                 } catch (error) {
@@ -159,14 +163,15 @@ phonesController.getPhonesJWT = async (req, res) => {
 
 phonesController.postCreatePhone = async (req, res) => {
     try {
-        const { phoneNumber, businessUnit, name, email } = req.body;
+        const { phoneNumber, businessUnit, name, email, user } = req.body;
         const newPhone = new phone({
             name: name,
             email: email,
             businessUnit: businessUnit,
             phoneNumber: phoneNumber,
             enable: false,
-            hits: 0
+            hits: 0,
+            user: user
         });
         await newPhone.save();
         res.status(201).json({
@@ -193,23 +198,43 @@ phonesController.postCreatePhoneJWT = async (req, res) => {
             const baererToken = baererHeader.split(" ")[1]
             req.token = baererToken;
             const permission = await permissionJWTVerify(baererToken, JWTFlag.permissionNameEdit)
-            if (permission == false) {
+            if (permission.flag == false || (
+                permission.user.user.permisions.find(permissionsAux => permissionsAux === 'admin') == undefined &&
+                permission.user.user.permisions.find(permissionsAux => permissionsAux === 'create') == undefined)) {
                 res.sendStatus(403)
             } else {
                 try {
-                    const { phoneNumber, businessUnit, name, email } = req.body;
-                    const newPhone = new phone({
-                        name: name,
-                        email: email,
-                        businessUnit: businessUnit,
-                        phoneNumber: phoneNumber,
-                        enable: false,
-                        hits: 0
-                    });
-                    await newPhone.save();
-                    res.status(201).json({
-                        mensaje: "Teléfono agregado",
-                    });
+                    if (permission.user.user.permisions.find(permissionsAux => permissionsAux === 'admin')) {
+                        const { phoneNumber, businessUnit, name, email, bussinesAccountId } = req.body;
+                        const newPhone = new phone({
+                            name: name,
+                            email: email,
+                            businessUnit: businessUnit,
+                            phoneNumber: phoneNumber,
+                            enable: false,
+                            hits: 0,
+                            bussinesAccountId: bussinesAccountId
+                        });
+                        await newPhone.save();
+                        res.status(201).json({
+                            mensaje: "Teléfono agregado",
+                        });
+                    } else {
+                        const { phoneNumber, businessUnit, name, email } = req.body;
+                        const newPhone = new phone({
+                            name: name,
+                            email: email,
+                            businessUnit: businessUnit,
+                            phoneNumber: phoneNumber,
+                            enable: false,
+                            hits: 0,
+                            bussinesAccountId: permission.user.user.bussinesAccountId
+                        });
+                        await newPhone.save();
+                        res.status(201).json({
+                            mensaje: "Teléfono agregado",
+                        });
+                    }
                 } catch (error) {
                     console.log(error);
                     if (error.code && error.code == 11000) {
@@ -242,7 +267,7 @@ phonesController.getPhoneHitsJWT = async (req, res) => {
             const baererToken = baererHeader.split(" ")[1]
             req.token = baererToken;
             const permission = await permissionJWTVerify(baererToken, JWTFlag.permissionNameView)
-            if (permission == false) {
+            if (permission.flag == false) {
                 res.sendStatus(403)
             } else {
                 try {
@@ -359,7 +384,9 @@ phonesController.putPhoneJWT = async (req, res) => {
             const baererToken = baererHeader.split(" ")[1]
             req.token = baererToken;
             const permission = await permissionJWTVerify(baererToken, JWTFlag.permissionNameEdit)
-            if (permission == false) {
+            if (permission.flag == false || (
+                permission.user.user.permisions.find(permissionsAux => permissionsAux === 'admin') == undefined &&
+                permission.user.user.permisions.find(permissionsAux => permissionsAux === 'edit') == undefined)) {
                 res.sendStatus(403)
             } else {
                 try {
@@ -438,29 +465,55 @@ phonesController.putPhoneByIdJWT = async (req, res) => {
             const baererToken = baererHeader.split(" ")[1]
             req.token = baererToken;
             const permission = await permissionJWTVerify(baererToken, JWTFlag.permissionNameEdit)
-            if (permission == false) {
+            if (permission.flag == false || (
+                permission.user.user.permisions.find(permissionsAux => permissionsAux === 'admin') == undefined &&
+                permission.user.user.permisions.find(permissionsAux => permissionsAux === 'edit') == undefined)) {
                 res.sendStatus(403)
             } else {
                 try {
-                    const { phoneNumber, businessUnit, name, email, enable } = req.body;
-                    const newPhone = {
-                        name: name,
-                        email: email,
-                        businessUnit: businessUnit,
-                        phoneNumber: phoneNumber,
-                        enable: enable,
-                        hits: 0
-                    };
-                    const updatePhone = await phone.findOneAndUpdate({ _id: req.params.id }, newPhone, { useFindAndModify: true });
-                    if (updatePhone == null) {
-                        res.status(201).json({
-                            mensaje: "Teléfono inexistente",
-                        });
+                    if (permission.user.user.permisions.find(permissionsAux => permissionsAux === 'admin')) {
+                        const { phoneNumber, businessUnit, name, email, enable, bussinesAccountId } = req.body;
+                        const newPhone = {
+                            name: name,
+                            email: email,
+                            businessUnit: businessUnit,
+                            phoneNumber: phoneNumber,
+                            enable: enable,
+                            hits: 0,
+                            bussinesAccountId: bussinesAccountId
+                        };
+                        const updatePhone = await phone.findOneAndUpdate({ _id: req.params.id, bussinesAccountId: permission.user.user.bussinesAccountId }, newPhone, { useFindAndModify: true });
+                        if (updatePhone == null) {
+                            res.status(201).json({
+                                mensaje: "Teléfono inexistente",
+                            });
+                        } else {
+                            res.status(201).json({
+                                mensaje: "Teléfono modificado",
+                            });
+                        }
                     } else {
-                        res.status(201).json({
-                            mensaje: "Teléfono modificado",
-                        });
+                        const { phoneNumber, businessUnit, name, email, enable } = req.body;
+                        const newPhone = {
+                            name: name,
+                            email: email,
+                            businessUnit: businessUnit,
+                            phoneNumber: phoneNumber,
+                            enable: enable,
+                            hits: 0
+                        };
+                        const updatePhone = await phone.findOneAndUpdate({ _id: req.params.id, bussinesAccountId: permission.user.user.bussinesAccountId }, newPhone, { useFindAndModify: true });
+                        if (updatePhone == null) {
+                            res.status(201).json({
+                                mensaje: "Teléfono inexistente",
+                            });
+                        } else {
+                            res.status(201).json({
+                                mensaje: "Teléfono modificado",
+                            });
+                        }
                     }
+
                 } catch (error) {
                     console.log(error);
                     res.status(500).json({
@@ -485,7 +538,7 @@ phonesController.deletePhoneJWT = async (req, res) => {
             const baererToken = baererHeader.split(" ")[1]
             req.token = baererToken;
             const permission = await permissionJWTVerify(baererToken, JWTFlag.permissionNameEdit)
-            if (permission == false) {
+            if (permission.flag == false) {
                 res.sendStatus(403)
             } else {
                 try {
@@ -546,19 +599,34 @@ phonesController.deletePhoneByIdJWT = async (req, res) => {
             const baererToken = baererHeader.split(" ")[1]
             req.token = baererToken;
             const permission = await permissionJWTVerify(baererToken, JWTFlag.permissionNameEdit)
-            if (permission == false) {
+            if (permission.flag == false || (
+                permission.user.user.permisions.find(permissionsAux => permissionsAux === 'admin') == undefined &&
+                permission.user.user.permisions.find(permissionsAux => permissionsAux === 'edit') == undefined)) {
                 res.sendStatus(403)
             } else {
                 try {
-                    const deletePhone = await phone.deleteOne({ _id: req.params.id });
-                    if (deletePhone.deletedCount > 0) {
-                        res.status(201).json({
-                            mensaje: "Teléfono eliminado",
-                        });
+                    if (permission.user.user.permisions.find(permissionsAux => permissionsAux === 'admin')) {
+                        const deletePhone = await phone.deleteOne({ _id: req.params.id });
+                        if (deletePhone.deletedCount > 0) {
+                            res.status(201).json({
+                                mensaje: "Teléfono eliminado",
+                            });
+                        } else {
+                            res.status(201).json({
+                                mensaje: "Teléfono no encontrado",
+                            });
+                        }
                     } else {
-                        res.status(201).json({
-                            mensaje: "Teléfono no encontrado",
-                        });
+                        const deletePhone = await phone.deleteOne({ _id: req.params.id, bussinesAccountId: permission.user.user.bussinesAccountId });
+                        if (deletePhone.deletedCount > 0) {
+                            res.status(201).json({
+                                mensaje: "Teléfono eliminado",
+                            });
+                        } else {
+                            res.status(201).json({
+                                mensaje: "Teléfono no encontrado",
+                            });
+                        }
                     }
                 } catch (error) {
                     console.log(error);
@@ -616,13 +684,21 @@ phonesController.getPhoneJWT = async (req, res) => {
             const baererToken = baererHeader.split(" ")[1]
             req.token = baererToken;
             const permission = await permissionJWTVerify(baererToken, JWTFlag.permissionNameView)
-            if (permission == false) {
+            if (permission.flag == false || (
+                permission.user.user.permisions.find(permissionsAux => permissionsAux === 'admin') == undefined &&
+                permission.user.user.permisions.find(permissionsAux => permissionsAux === 'view') == undefined)) {
                 res.sendStatus(403)
             } else {
                 try {
-                    const id = req.params.id
-                    const objectPhone = await phone.findById(id);
-                    res.status(200).json(objectPhone);
+                    if (permission.user.user.permisions.find(permissionsAux => permissionsAux === 'admin')) {
+                        const id = req.params.id
+                        const objectPhone = await phone.findById(id);
+                        res.status(200).json(objectPhone);
+                    } else {
+                        const id = req.params.id
+                        const objectPhone = await phone.find({_id: id, bussinesAccountId: permission.user.user.bussinesAccountId});
+                        res.status(200).json(objectPhone);
+                    }
                 } catch (error) {
                     console.log(error);
                     res.status(500).json({
